Refactor: Application of shadcn component Part 3: - Notes application structure by separating components and implementing FiltersSidebar and NoteDialog. - Migrate note handling logic to NotesPage, enhancing state management and improving code organization.
